refactor: change directory/name
