+++ src/components/ui/scroll-to-top.tsx
@@ -1,5 +1,5 @@
 import { useState, useEffect } from "react";
-import { Button } from "./ui/button";
+import { Button } from "./button";
 import { ArrowUpToLine } from "lucide-react";
 
 export const ScrollToTop = () => {
@@ -27,10 +27,10 @@ export const ScrollToTop = () => {
       {showTopBtn && (
         <Button
           onClick={goToTop}
-          className="fixed bottom-4 right-4 opacity-90 shadow-md"
+          className="bg-wppButton fixed bottom-4 right-4"
           size="icon"
         >
-          <ArrowUpToLine className="h-4 w-4" />
+          <ArrowUpToLine className="text-black h-4 w-4" />
         </Button>
       )}
     </>
